+++ src/feature_writer.rs
@@ -7,7 +7,8 @@ use std::io::Write;
 #[derive(PartialEq)]
 enum State {
     New,
-    Started,
+    WritingFeatures,
+    WritingForeignMembers,
     Finished,
 }
 
@@ -23,6 +24,9 @@ impl<W: Write> FeatureWriter<W> {
     /// To append features from your custom structs, use [`FeatureWriter::serialize`].
     ///
     /// To append features from [`Feature`] use [`FeatureWriter::write_feature`].
+    ///
+    /// To write a foreign member, use [`FeatureWriter::write_foreign_member`] before appending any
+    /// features.
     pub fn from_writer(writer: W) -> Self {
         Self {
             writer,
@@ -41,10 +45,14 @@ impl<W: Write> FeatureWriter<W> {
             }
             State::New => {
                 self.write_prefix()?;
-                self.state = State::Started;
+                self.state = State::WritingFeatures;
             }
-            State::Started => {
+            State::WritingFeatures => {
                 self.write_str(",")?;
+            }
+            State::WritingForeignMembers => {
+                self.write_str(r#" "features": ["#)?;
+                self.state = State::WritingFeatures;
             }
         }
         serde_json::to_writer(&mut self.writer, feature)?;
@@ -158,13 +166,49 @@ impl<W: Write> FeatureWriter<W> {
             }
             State::New => {
                 self.write_prefix()?;
-                self.state = State::Started;
+                self.state = State::WritingFeatures;
             }
-            State::Started => {
+            State::WritingFeatures => {
                 self.write_str(",")?;
+            }
+            State::WritingForeignMembers => {
+                self.write_str(r#" "features": ["#)?;
+                self.state = State::WritingFeatures;
             }
         }
         to_feature_writer(&mut self.writer, value)
+    }
+
+    /// Write a [foreign member](https://datatracker.ietf.org/doc/html/rfc7946#section-6) to the
+    /// output stream. This must be done before appending any features.
+    pub fn write_foreign_member<T: ?Sized + Serialize>(
+        &mut self,
+        key: &str,
+        value: &T,
+    ) -> Result<()> {
+        match self.state {
+            State::Finished => Err(Error::InvalidWriterState(
+                "cannot write foreign member when writer has already finished",
+            )),
+            State::New => {
+                self.write_str(r#"{ "type": "FeatureCollection", "#)?;
+                write!(self.writer, "\"{key}\": ")?;
+                serde_json::to_writer(&mut self.writer, value)?;
+                self.write_str(",")?;
+
+                self.state = State::WritingForeignMembers;
+                Ok(())
+            }
+            State::WritingFeatures => Err(Error::InvalidWriterState(
+                "must write foreign members before any features",
+            )),
+            State::WritingForeignMembers => {
+                write!(self.writer, "\"{key}\": ")?;
+                serde_json::to_writer(&mut self.writer, value)?;
+                self.write_str(",")?;
+                Ok(())
+            }
+        }
     }
 
     /// Writes the closing syntax for the FeatureCollection.
@@ -183,7 +227,7 @@ impl<W: Write> FeatureWriter<W> {
                 self.write_prefix()?;
                 self.write_suffix()?;
             }
-            State::Started => {
+            State::WritingFeatures | State::WritingForeignMembers => {
                 self.state = State::Finished;
                 self.write_suffix()?;
             }
@@ -374,6 +418,58 @@ mod tests {
             ]
         });
 
+        let actual_json: JsonValue = serde_json::from_slice(&buffer).expect("valid json");
+        assert_eq!(actual_json, expected)
+    }
+
+    #[test]
+    fn write_foreign_members() {
+        let mut buffer: Vec<u8> = vec![];
+        {
+            let mut writer = FeatureWriter::from_writer(&mut buffer);
+
+            writer.write_foreign_member("extra", "string").unwrap();
+            writer.write_foreign_member("list", &[1, 2, 3]).unwrap();
+            writer
+                .write_foreign_member("nested", &json!({"foo": "bar"}))
+                .unwrap();
+
+            let record_1 = {
+                let mut props = serde_json::Map::new();
+                props.insert("name".to_string(), "Mishka".into());
+                props.insert("age".to_string(), 12.into());
+
+                Feature {
+                    bbox: None,
+                    geometry: Some(crate::Geometry::from(crate::Value::Point(vec![1.1, 1.2]))),
+                    id: None,
+                    properties: Some(props),
+                    foreign_members: None,
+                }
+            };
+
+            writer.write_feature(&record_1).unwrap();
+            writer.flush().unwrap();
+        }
+
+        let expected = json!({
+            "type": "FeatureCollection",
+            "extra": "string",
+            "list": [1, 2, 3],
+            "nested": {
+                "foo": "bar",
+            },
+            "features": [
+                {
+                    "type": "Feature",
+                    "geometry": { "type": "Point", "coordinates": [1.1, 1.2] },
+                    "properties": { "name": "Mishka", "age": 12
+                    }
+                },
+            ]
+        });
+
+        println!("{}", String::from_utf8(buffer.clone()).unwrap());
         let actual_json: JsonValue = serde_json::from_slice(&buffer).expect("valid json");
         assert_eq!(actual_json, expected)
     }