+++ tests/roundtrip.rs
@@ -47,7 +47,7 @@ mod roundtrip_tests {
     /// Verifies that we can parse and then re-encode geojson back to the same representation
     /// without losing any data.
     fn test_round_trip(file_path: &str) {
-        let mut file = File::open(&file_path).unwrap();
+        let mut file = File::open(file_path).unwrap();
         let mut file_contents = String::new();
         let _ = file.read_to_string(&mut file_contents);